star-trails/requirements.txt
numpy==2.0.2
pillow==11.2.1
tqdm==4.67.1